requirements.txt
alembic==1.11.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bcolz-zipline==1.2.6
beautifulsoup4==4.12.2
blosc2==2.0.0
Bottleneck==1.3.7
bt==0.2.9
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
contourpy==1.1.0
cvxpy==1.3.2
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
ecos==2.0.12
empyrical==0.5.5
empyrical-reloaded==0.5.9
exceptiongroup==1.1.2
exchange-calendars==4.2.8
executing==1.2.0
ffn==0.3.7
fonttools==4.41.1
frozendict==2.3.8
greenlet==2.0.2
h5py==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
intervaltree==3.1.0
ipykernel==6.24.0
ipython==8.12.2
iso3166==2.1.1
iso4217==1.11.20220401
jedi==0.18.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
lru-dict==1.2.0
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multipledispatch==1.0.0
multitasking==0.0.11
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.24.4
osqp==0.6.3
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyfolio-reloaded==0.9.5
Pygments==2.15.1
pyluach==2.2.0
pyparsing==3.0.9
pyportfolioopt==1.5.5
PyPrind==2.11.3
pyrate-limiter==2.10.0
python-coinmarketcap==0.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-interface==1.6.1
pytz==2023.3
pyzmq==25.1.0
qdldl==0.1.7.post0
requests==2.31.0
requests-cache==1.1.0
requests-ratelimiter==0.4.0
scikit-learn==1.3.0
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
statsmodels==0.14.0
TA-Lib==0.4.27
tables==3.8.0
tabulate==0.9.0
threadpoolctl==3.2.0
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
yfinance==0.2.26
zipline-reloaded==3.0
zipp==3.16.2
